+++ packages/app/src/app/components/Footer.tsx
@@ -3,10 +3,12 @@ import { FC } from "react";
 import { makeStyles } from "tss-react/mui";
 import { ReactComponent as BackgroundSvg } from "../../assets/bg.svg";
 import oxLogo from "../../assets/ox.svg";
+import { ShareButton } from "./ShareButton";
 import { ToggleThemeButton } from "./ToggleThemeButton";
 
 export const Footer: FC = () => {
   const { classes } = useStyles();
+  
 
   return (
     <div className={classes.footer}>
@@ -45,7 +47,10 @@ export const Footer: FC = () => {
               MegaLinter
             </a>
           </Typography>
-          <ToggleThemeButton />
+          <div className={classes.actions}>
+            <ShareButton />
+            <ToggleThemeButton />
+          </div>
         </div>
       </Container>
       <BackgroundSvg />
@@ -96,5 +101,9 @@ const useStyles = makeStyles()((theme: Theme) => ({
   oxLogo: {
     width: "1.3em",
     height: "1.3em",
+  },
+  actions: {
+    display: "flex",
+    gap: theme.spacing(2),
   },
 }));